Rework enemies spawning & animations
