+++ app/canvas/index.ts
@@ -1,8 +1,8 @@
-export * from './allies'
+export * from './ally'
 export * from './animations'
 export * from './camera'
 export * from './constants'
-export * from './enemies'
+export * from './enemy'
 export * from './ground'
 export * from './light'
 export * from './renderer'
